устранения недостатков
