+++ src/main/java/sky/group/homeworkgroup/controller/RuleController.java
@@ -5,8 +5,8 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import sky.group.homeworkgroup.model_dinamicbase.Rule;
-import sky.group.homeworkgroup.serviceRule.DinamicService;
+import sky.group.homeworkgroup.model.model_dinamicbase.Rule;
+import sky.group.homeworkgroup.service.request.RuleService;
 
 import java.util.List;
 
@@ -14,17 +14,16 @@ import java.util.List;
 @RestController
 @RequestMapping("/rule")
 public class RuleController {
-    private final DinamicService dinamicService;
+    private final RuleService ruleService;
 
-    public RuleController(DinamicService dinamicService) {
-        this.dinamicService = dinamicService;
+    public RuleController(RuleService ruleService) {
+        this.ruleService = ruleService;
     }
-
 
     @DeleteMapping("/delete")
     @Operation(summary = "Проводим удаление рекомендаций с id совета из продукта по заданному id продукта")
     public ResponseEntity<String> deleteRule(@RequestParam("ID product") Long idProduct, @RequestParam("ID rule") Long idRule) {
-        String result = dinamicService.deleteRule(idProduct, idRule);
+        String result = ruleService.deleteRule(idProduct, idRule);
         if (result.equals("Строка удалена")) {
             return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
         } else {
@@ -35,9 +34,8 @@ public class RuleController {
     @GetMapping("/allRule")
     @Operation(summary = "Отображаем имеющиеся советы по продукту с заданным id")
     public List<Rule> AdviceAll(@RequestParam("ID product") Long idProduct) {
-        return dinamicService.allAdvice(idProduct);
+        return ruleService.allAdvice(idProduct);
     }
-
 }
