+++ src/main/java/sky/group/homeworkgroup/controller/SuggestionsForClientController.java
@@ -7,7 +7,8 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import sky.group.homeworkgroup.model.OutputData;
-import sky.group.homeworkgroup.service.ServiceClient;
+import sky.group.homeworkgroup.service.ClientService;
+import sky.group.homeworkgroup.service.DinamicClientService;
 
 import java.util.List;
 import java.util.Optional;
@@ -16,15 +17,23 @@ import java.util.UUID;
 @RestController
 @RequestMapping("/recommendation")
 public class SuggestionsForClientController {
-    private final ServiceClient serviceClients;
+    private final ClientService serviceClients;
+    private final DinamicClientService dinamicClientService;
 
-    public SuggestionsForClientController(ServiceClient serviceClients) {
+    public SuggestionsForClientController(ClientService serviceClients, DinamicClientService dinamicClientService) {
         this.serviceClients = serviceClients;
+        this.dinamicClientService = dinamicClientService;
     }
 
     @GetMapping("{user_Id}")
     @Operation(summary = "Вводим id клиента")
     public ResponseEntity<Optional<List<OutputData>>> getRecommendations(@PathVariable UUID user_Id) {
         return ResponseEntity.ok(Optional.of(serviceClients.searchForRecommendations(user_Id)));
+    }
+
+    @GetMapping("/dynamic/{user_Id}")
+    @Operation(summary = "Вводим id клиента для динамического анализа")
+    public ResponseEntity<Optional<List<OutputData>>> getRecommendationsDinamic(@PathVariable UUID user_Id) {
+        return ResponseEntity.ok(Optional.of(dinamicClientService.searchForRecommendationsDinamic(user_Id)));
     }
 }
